refactor: Consolidate documentation structure and simplify routing
This commit introduces a significant restructuring of the documentation system:
- Simplified source configuration by removing individual source configurations
- Unified documentation routing under a single `/docs` path
- Removed separate layouts and pages for different documentation sections
- Consolidated content into a single source with a new meta.json
- Updated global styles and layout configuration
- Removed privacy regulations and release notes content

The changes streamline the documentation architecture, making it easier to maintain and navigate while reducing code complexity.
