--- apps/docs/src/app/docs/privacy-regulations/page.tsx
@@ -1,5 +1,0 @@
-import { Laws } from './_components/laws';
-
-export default function Page() {
-	return <Laws />;
-}
