+++ apps/docs/src/app/not-found.tsx
@@ -1,11 +1,8 @@
-import Navbar from '~/components/docs/navbar';
 import NotFoundClient from './not-found.client';
-
 export default function NotFound() {
 	return (
-		<>
-			<Navbar />
+		<div className="not-found-layout">
 			<NotFoundClient />
-		</>
+		</div>
 	);
 }
