--- apps/docs/src/content/privacy-regulations/africa/popia.mdx
@@ -1,106 +1,0 @@
----
-title: South Africa's Protection of Personal Information Act (POPIA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in South Africa.
----
-
-## Why POPIA Matters for Your Startup
-
-If your startup processes personal information of South African residents, adhering to the Protection of Personal Information Act (POPIA) is crucial. Enacted to safeguard individuals' personal data, POPIA imposes specific obligations on businesses operating within or interacting with individuals in South Africa.
-
-## Key Provisions of POPIA
-
-### 1. Accountability
-
-**Requirement:** Ensure compliance with POPIA's principles throughout your organization.
-
-**Action Steps:**
-- Assign a dedicated individual or team to oversee data protection efforts.
-- Develop and implement data protection policies and procedures.
-- Regularly review and update compliance measures.
-
-### 2. Processing Limitation
-
-**Requirement:** Process personal information lawfully and minimally.
-
-**Action Steps:**
-- Collect only data that is necessary for your specified purposes.
-- Obtain explicit consent from individuals before processing their data.
-- Allow individuals to withdraw consent easily.
-
-### 3. Purpose Specification
-
-**Requirement:** Collect personal information for specific, explicitly defined, and legitimate reasons.
-
-**Action Steps:**
-- Clearly define and document the purposes for data collection.
-- Inform individuals about the specific purposes for which their data is being collected.
-- Retain data only as long as necessary to fulfill these purposes.
-
-### 4. Further Processing Limitation
-
-**Requirement:** Ensure that any further processing of personal information is compatible with the original purpose of collection.
-
-**Action Steps:**
-- Evaluate whether new processing activities align with the original purposes.
-- Obtain additional consent if processing for a new purpose.
-
-### 5. Information Quality
-
-**Requirement:** Maintain accurate, complete, and up-to-date personal information.
-
-**Action Steps:**
-- Implement procedures to verify and update data regularly.
-- Encourage individuals to inform you of any changes to their personal information.
-
-### 6. Openness
-
-**Requirement:** Be transparent about data collection and processing practices.
-
-**Action Steps:**
-- Develop a clear and accessible privacy policy detailing your data practices.
-- Inform individuals about the types of personal information collected and the purposes thereof.
-
-### 7. Security Safeguards
-
-**Requirement:** Protect personal information against risks such as loss, unauthorized access, and destruction.
-
-**Action Steps:**
-- Implement appropriate technical and organizational security measures.
-- Conduct regular security assessments and audits.
-- Train employees on data security best practices.
-
-### 8. Data Subject Participation
-
-**Requirement:** Respect individuals' rights to access, correct, and delete their personal information.
-
-**Action Steps:**
-- Establish procedures to handle data access, correction, and deletion requests.
-- Respond to such requests promptly and in accordance with legal requirements.
-
-## Penalties for Non-Compliance
-
-Non-compliance with POPIA can lead to severe consequences:
-
-- **Administrative Fines:** Up to ZAR 10 million.
-- **Criminal Charges:** Including imprisonment for serious offenses.
-- **Reputational Damage:** Loss of customer trust and potential business opportunities.
-
-## Compliance Roadmap for Startups
-
-**Phase 1: Assessment**
-- Conduct a comprehensive audit of your data processing activities.
-- Identify gaps between current practices and POPIA requirements.
-
-**Phase 2: Implementation**
-- Develop and implement policies and procedures to address identified gaps.
-- Train employees on POPIA requirements and data protection best practices.
-
-**Phase 3: Monitoring and Maintenance**
-- Regularly review and update data protection practices.
-- Stay informed about legal developments and adjust compliance efforts accordingly.
-
-## Final Thoughts
-
-Complying with POPIA is not only a legal obligation but also an opportunity to build trust with your customers by demonstrating a commitment to protecting their personal information. By proactively addressing these requirements, your startup can mitigate risks and establish a strong foundation for responsible data management.
-
-For detailed guidance and resources, consider consulting with privacy experts or utilizing compliance tools designed to assist businesses in navigating POPIA's requirements.
