--- apps/docs/src/content/privacy-regulations/asia/singapore-pdpa.mdx
@@ -1,46 +1,0 @@
----
-title: Singapore's Personal Data Protection Act (PDPA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Singapore.
----
-
-## Overview of the Personal Data Protection Act (PDPA)
-
-nacted in 2012, Singapore's Personal Data Protection Act (PDPA) establishes a comprehensive framework governing the collection, use, disclosure, and care of personal data.t balances individuals' rights to protect their personal data with organizations' needs to collect and use such data for legitimate and reasonable purposes.
-## Key Definitions
-
-- **Personal Data**: ata, whether true or not, about an individual who can be identified from that data, or from that data and other information to which the organization has or is likely to have access.
-- **Organization**: ny individual, company, association, or body of persons, corporate or unincorporated, whether or not formed or recognized under the law of Singapore, or resident, or having an office or a place of business, in Singapore.
-## Data Protection Obligations
-
-he PDPA outlines several obligations for organizations handling personal data:
-1. **Consent Obligation**: Organizations must obtain an individual's consent before collecting, using, or disclosing their personal data, unless an exception applies.
-2. **Purpose Limitation Obligation**: Personal data can only be collected, used, or disclosed for purposes that a reasonable person would consider appropriate under the circumstances.
-3. **Notification Obligation**: Organizations must inform individuals of the purposes for which their personal data will be collected, used, or disclosed.
-4. **Access and Correction Obligation**: Upon request, organizations must provide individuals with access to their personal data and allow them to correct any inaccuracies.
-5. **Accuracy Obligation**: Organizations must make reasonable efforts to ensure that personal data collected is accurate and complete if it is likely to be used to make a decision affecting the individual or disclosed to another organization.
-6. **Protection Obligation**: Organizations must implement reasonable security arrangements to protect personal data from unauthorized access, collection, use, disclosure, copying, modification, disposal, or similar risks.
-7. **Retention Limitation Obligation**: Personal data should not be retained longer than necessary for legal or business purposes.
-8. **Transfer Limitation Obligation**: Organizations transferring personal data outside of Singapore must ensure that the receiving organization provides a standard of protection comparable to the PDPA.
-9. **Data Breach Notification Obligation**: Organizations are required to notify the Personal Data Protection Commission (PDPC) and affected individuals of data breaches that result in, or are likely to result in, significant harm to individuals.
-10. **Accountability Obligation**: Organizations must implement policies and practices to meet their obligations under the PDPA and make information about their data protection policies available upon request.
-
-## Rights of Individuals
-
-nder the PDPA, individuals have the right to:
-- **Access**: equest access to their personal data held by an organization.
-- **Correction**: equest correction of inaccuracies in their personal data.
-- **Withdrawal of Consent**: ithdraw consent for the collection, use, or disclosure of their personal data, subject to legal or contractual restrictions.
-## Enforcement and Penalties
-
-he PDPC is responsible for enforcing the PDPA. Organizations that fail to comply may face financial penalties, directions to cease non-compliant activities, or other corrective measures.
-## Preparing for PDPA Compliance
-
-Organizations should take the following steps to comply with the PDPA:
-1. **Assess Applicability**: Determine if the PDPA applies to your organization's data processing activities.
-2. **Data Inventory**: Conduct a thorough inventory of personal data collected, used, and disclosed by your organization.
-3. **Policy Development**: Develop and implement data protection policies and practices in line with PDPA requirements.
-4. **Employee Training**: Train employees on PDPA obligations and your organization's data protection policies.
-5. **Consent Management**: Establish processes to obtain, manage, and document individuals' consent for data collection, use, and disclosure.
-6. **Data Breach Response Plan**: Develop a response plan for data breaches, including notification procedures to the PDPC and affected individuals.
-7. **Regular Audits**: Conduct regular audits to ensure ongoing compliance with the PDPA.
-By implementing these measures, organizations can align with the PDPA's requirements and uphold individuals' personal data protection rights in Singapore.
