--- apps/docs/src/content/privacy-regulations/asia/thailand-pdpa.mdx
@@ -1,71 +1,0 @@
----
-title: Thailand's Personal Data Protection Act (PDPA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Thailand.
----
-
-The Personal Data Protection Act (PDPA), B.E. 2562 (2019), is Thailand's first comprehensive data protection legislation, which came into effect on June 1, 2022. The PDPA aligns closely with international standards, such as the European Union's General Data Protection Regulation (GDPR), aiming to safeguard personal data and uphold individuals' privacy rights.citeturn0search9
-
-## Applicability of the PDPA
-
-The PDPA applies to any organization, regardless of its location, that collects, uses, or discloses personal data of individuals residing in Thailand. This includes both data controllers and data processors operating within or outside Thailand, provided their activities involve offering goods or services to individuals in Thailand or monitoring their behavior within the country.citeturn0search13
-
-## Key Definitions
-
-- **Personal Data**: Any information relating to an identified or identifiable natural person, directly or indirectly, excluding data of deceased individuals.
-- **Data Controller**: A person or entity with the authority to make decisions regarding the collection, use, or disclosure of personal data.
-- **Data Processor**: A person or entity that processes personal data on behalf of the data controller.
-
-## Core Principles of the PDPA
-
-The PDPA establishes several fundamental principles for personal data processing:
-1. **Lawful and Fair Processing**: Personal data must be processed lawfully and fairly, ensuring transparency with data subjects.
-2. **Purpose Limitation**: Data should be collected for specified, explicit, and legitimate purposes and not further processed in a manner incompatible with those purposes.
-3. **Data Minimization**: Only data necessary for the intended purposes should be collected and processed.
-4. **Accuracy**: Personal data must be accurate, complete, and kept up-to-date to prevent misunderstandings or harm.
-5. **Storage Limitation**: Data should be retained only as long as necessary for the purposes for which it was collected.
-6. **Integrity and Confidentiality**: Appropriate security measures must be implemented to protect personal data against unauthorized or unlawful processing, accidental loss, destruction, or damage.
-7. **Accountability**: Data controllers are responsible for ensuring compliance with the PDPA and must be able to demonstrate such compliance.
-
-## Rights of Data Subjects
-
-The PDPA grants individuals several rights concerning their personal data:
-- **Right to Access**: Individuals can request access to their personal data held by an organization and obtain information about how it is processed.
-- **Right to Rectification**: Individuals can request correction of inaccurate or incomplete personal data.
-- **Right to Erasure (Right to be Forgotten)**: Individuals can request deletion of their personal data under certain circumstances, such as when the data is no longer necessary for the purposes it was collected.
-- **Right to Restriction of Processing**: Individuals can request the limitation of their data processing under specific conditions.
-- **Right to Data Portability**: Individuals have the right to receive their personal data in a structured, commonly used, and machine-readable format and transmit it to another data controller.
-- **Right to Object**: Individuals can object to the processing of their personal data, including for direct marketing purposes.
-- **Right to Withdraw Consent**: Individuals can withdraw their consent for data processing at any time, without affecting the lawfulness of processing based on consent before its withdrawal.
-- **Right to Lodge a Complaint**: Individuals can file complaints with the Personal Data Protection Committee (PDPC) if they believe their data protection rights have been violated.
-
-## Obligations of Data Controllers and Processors
-
-Organizations handling personal data have specific responsibilities under the PDPA:
-- **Consent Acquisition**: Obtain explicit consent from data subjects before collecting, using, or disclosing their personal data, unless an exemption applies.
-- **Data Protection Officer (DPO)**: Appoint a DPO if the organization engages in large-scale data processing, processes sensitive data, or is a public authority.
-- **Data Processing Records**: Maintain records of data processing activities, including purposes, data categories, and recipient details.
-- **Data Breach Notification**: Notify the PDPC of any data breaches within 72 hours of becoming aware, and inform affected data subjects if the breach poses a high risk to their rights and freedoms.
-- **Data Protection Impact Assessments (DPIAs)**: Conduct DPIAs for processing activities that may result in high risks to individuals' rights and freedoms.
-- **Cross-Border Data Transfers**: Ensure that personal data transferred outside Thailand is afforded adequate protection, comparable to the PDPA's standards.
-
-## Penalties for Non-Compliance
-
-Non-compliance with the PDPA can result in significant penalties:
-- **Administrative Fines**: Up to THB 5 million, depending on the severity of the violation.
-- **Criminal Penalties**: In severe cases, imprisonment of up to one year and/or fines up to THB 1 million.
-- **Civil Liabilities**: Data subjects may claim compensation for damages resulting from PDPA violations.
-
-## Steps to Achieve PDPA Compliance
-
-Organizations should take the following steps to comply with the PDPA:
-
-1. **Data Mapping**: Identify and document the personal data collected, processed, and stored by the organization.
-2. **Policy Development**: Establish or update privacy policies to align with PDPA requirements, clearly outlining data processing activities and purposes.
-3. **Consent Management**: Implement mechanisms to obtain, record, and manage data subjects' consent effectively.
-4. **Data Subject Rights**: Develop procedures to facilitate and respond to data subjects' rights requests promptly.
-5. **Security Measures**: Implement appropriate technical and organizational security measures to protect personal data.
-6. **Third-Party Management**: Ensure that any third-party service providers who process personal data on your behalf comply with PDPA requirements through due diligence, data processing agreements, and monitoring.
-7. **Data Breach Response Plan**: Develop and implement a comprehensive plan for identifying, containing, investigating and notifying relevant parties about data breaches.
-8. **Regular Audits and Monitoring**: Conduct regular audits of data processing activities to ensure ongoing PDPA compliance and identify areas for improvement.
-9. **Documentation and Record-Keeping**: Maintain comprehensive records of all data processing activities, consent records, and data protection impact assessments.
-10. **Stay Informed on Regulatory Updates**: Keep abreast of changes to the PDPA and related regulations through industry engagement and legal consultation.
