--- apps/docs/src/content/privacy-regulations/canada/pipeda.mdx
@@ -1,151 +1,0 @@
----
-title: Personal Information Protection and Electronic Documents Act (PIPEDA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Canada.
----
-## Why You Need to Care
-
-PIPEDA is Canada's federal privacy law. If your product collects data from Canadian users, you need to comply. Unlike GDPR or CCPA, PIPEDA focuses on reasonable data handling rather than strict technical requirements.
-
-## When PIPEDA Applies
-
-Your product needs to comply if it:
-- Handles personal information in commercial activities in Canada
-- Transfers data across provincial or national borders 
-- Operates outside Quebec, British Columbia, and Alberta (these provinces have their own laws)
-
-Even if you're not based in Canada, PIPEDA applies when handling Canadian users' data.
-
-## What Counts as Personal Information
-
-PIPEDA covers any data that could identify a user:
-
-Standard Digital Data:
-- Email addresses and usernames
-- IP addresses and device IDs
-- Cookie data
-- Account details
-- Location data
-- Browser and device info
-- Usage analytics
-- Authentication records
-
-User-Created Data:
-- Uploaded content
-- Communications
-- Profile information
-- Usage patterns
-- Preferences
-
-Financial Data:
-- Payment details
-- Transaction history
-- Subscription info
-- Billing addresses
-
-## Core Requirements
-
-### 1. Get Real Consent
-
-Users must understand and agree to data collection:
-- Explain what you're collecting and why
-- Use clear language
-- Make consent active (no pre-checked boxes)
-- Let users change their mind
-- Keep consent records
-
-### 2. Only Collect What You Need
-
-Example: An e-commerce app needs shipping addresses but might not need phone numbers for every order. If you want extra data, have a clear reason and explain it.
-
-### 3. Keep Data Safe
-
-Basic security requirements:
-- Encrypt data (in transit and stored)
-- Use strong authentication
-- Control who can access what
-- Watch for unauthorized access
-- Keep security logs
-- Check security regularly
-
-### 4. Give Users Control
-
-Build features that let users:
-- See their data
-- Fix mistakes
-- Download their info
-- Delete their data
-- Understand how you use their info
-- Complain if needed
-
-## Privacy Policy Must-Haves
-
-Explain in plain language:
-- What you collect
-- Why you collect it
-- How you use it
-- Who you share it with
-- How users can control their data
-- How you protect information
-- How long you keep data
-- Who to contact about privacy
-
-## Working with Third-Party Services
-
-When using external services:
-- Check their privacy practices
-- Get proper agreements
-- Monitor what they do
-- Document data sharing
-- Review access regularly
-- Keep a list of all vendors
-
-## Common Problems to Watch For
-
-### Analytics
-- Use privacy-friendly analytics when possible
-- Get consent before tracking
-- Only track what you need
-- Document why you need each metric
-
-### Authentication
-- Verify identity securely
-- Don't ask for extra personal info
-- Protect login credentials
-- Watch for suspicious activity
-
-### Data Storage
-- Choose storage locations carefully
-- Encrypt the data
-- Clean up old data
-- Keep backups secure
-- Log who accesses what
-
-## Handling Data Breaches
-
-If something goes wrong:
-1. Figure out what happened
-2. Stop the breach
-3. Assess the risk
-4. Tell affected users if needed
-5. Tell the privacy commissioner if serious
-6. Keep records
-7. Fix the problem
-
-## Looking Ahead
-
-Stay ready for:
-- AI privacy issues
-- Changes in data transfer rules
-- New privacy laws
-- Technology changes
-- User expectations
-
-## When to Get Expert Help
-
-Consider legal advice when:
-- Handling sensitive data
-- Expanding to new markets
-- After security problems
-- For complex data sharing
-
-Keep your privacy practices simple and document everything. Focus on being transparent with users and only collecting data you actually need.
