--- apps/docs/src/content/privacy-regulations/canada/quebec-law-25.mdx
@@ -1,115 +1,0 @@
----
-title: Quebec Law 25
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Quebec.
----
-
-## Understanding Quebec's Law 25
-
-Quebec's Law 25, formerly known as Bill 64, represents a significant update to privacy regulations in Quebec. Proposed by the Quebec government in June 2020, it became law in September 2021. This legislation introduces new rules that businesses in Quebec must adhere to, with changes phased in over three years starting in 2022.
-
-## Key Provisions of Law 25
-
-### 1. Appointment of a Privacy Officer
-
-**Requirement:** Designate a person responsible for the protection of personal information within your organization. This individual oversees compliance and reports to the highest authority in the company.
-
-**Action Steps:**
-- Assign a qualified individual as the Privacy Officer.
-- Ensure they have a direct reporting line to top management.
-- Publicize the officer's contact information to facilitate communication.
-
-### 2. Privacy Impact Assessments (PIAs)
-
-**Requirement:** Conduct PIAs for any project involving the collection, use, or disclosure of personal information.
-
-**Action Steps:**
-- Develop a PIA process to evaluate privacy risks.
-- Document assessments and implement measures to mitigate identified risks.
-
-### 3. Consent and Transparency
-
-**Requirement:** Obtain clear and informed consent from individuals before collecting, using, or disclosing their personal information.
-
-**Action Steps:**
-- Implement consent mechanisms that are explicit and specific.
-- Provide individuals with clear information about data collection purposes and usage.
-
-### 4. Data Subject Rights
-
-**Requirement:** Respect individuals' rights to access, correct, and delete their personal information.
-
-**Action Steps:**
-- Establish procedures to handle requests for data access, correction, and deletion.
-- Ensure responses are timely and comply with legal requirements.
-
-### 5. Data Breach Notification
-
-**Requirement:** Notify the Commission d'accès à l'information (CAI) and affected individuals of any confidentiality incidents that present a risk of serious injury.
-
-**Action Steps:**
-- Develop a breach response plan outlining notification procedures.
-- Train employees to recognize and report data breaches promptly.
-
-### 6. Data Minimization and Retention
-
-**Requirement:** Collect only necessary personal information and retain it only for as long as needed to fulfill the specified purposes.
-
-**Action Steps:**
-- Review data collection practices to ensure they align with the principle of data minimization.
-- Establish and enforce data retention and deletion policies.
-
-## Penalties for Non-Compliance
-
-Non-compliance with Law 25 can result in substantial penalties:
-
-- **Administrative Fines:** Up to $10 million CAD or 2% of worldwide turnover, whichever is higher.
-- **Penal Fines:** For severe violations, fines can reach $25 million CAD or 4% of worldwide turnover.
-
-Additionally, public disclosure of violations can damage your startup's reputation and erode customer trust.
-
-## Compliance Roadmap for Startups and Developers
-
-**Phase 1: Assessment**
-- Conduct a comprehensive audit of personal information handling practices.
-- Identify gaps between current practices and Law 25 requirements.
-
-**Phase 2: Implementation**
-- Develop and implement policies and procedures to address identified gaps.
-- Train employees on new policies and the importance of data protection.
-
-**Phase 3: Monitoring and Maintenance**
-- Regularly review and update data protection practices.
-- Stay informed about legal developments and adjust compliance efforts accordingly.
-
-## Practical Steps for Developers
-
-### Data Mapping
-
-- **Inventory Personal Data:** Identify all personal data collected, processed, and stored by your organization.
-- **Assess Data Flow:** Understand how data moves within your organization and to external parties.
-- **Categorize Data:** Classify data based on sensitivity and processing purposes.
-
-### Technical Implementation
-
-- **Privacy by Design:** Integrate privacy considerations into the development of products and services from the outset.
-- **Security Measures:** Implement robust security protocols, including encryption, access controls, and regular vulnerability assessments.
-- **User Rights Implementation:** Develop functionalities that allow users to easily exercise their rights under Law 25.
-- **Consent Management:** Establish clear mechanisms for obtaining, recording, and managing user consent for data processing activities.
-
-### Documentation
-
-- **Privacy Policy:** Create a transparent and comprehensive privacy policy that outlines your data processing activities.
-- **Data Processing Records:** Maintain detailed records of all data processing activities.
-- **Security Measures:** Document the technical and organizational measures implemented to protect personal data.
-- **Incident Response Procedures:** Develop and document procedures for responding to data breaches.
-
-### Ongoing Maintenance
-
-- **Regular Reviews:** Periodically assess your data processing activities and compliance measures.
-- **Update Documentation:** Keep all compliance-related documentation current.
-- **Train Team Members:** Provide ongoing training to employees on Law 25 requirements.
-- **Monitor Compliance:** Continuously monitor your organization's adherence to Law 25 requirements.
-
-## Final Thoughts
-
-Achieving compliance with Quebec's Law 25 is not only a legal obligation but also an opportunity to build trust with your customers by demonstrating a commitment to protecting their personal information. By proactively addressing these requirements, your startup can mitigate risks and establish a strong foundation for responsible data management.
