--- apps/docs/src/content/privacy-regulations/europe/gdpr-uk.mdx
@@ -1,101 +1,0 @@
----
-title: UK General Data Protection Regulation (UK GDPR)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in the UK.
----
-
-## What You Need to Know About UK GDPR
-
-As the United Kingdom has exited the European Union, the General Data Protection Regulation (GDPR) has been incorporated into UK law as the UK GDPR, complemented by the Data Protection Act 2018. For startups and developers operating within the UK, it's essential to understand and implement the UK GDPR's requirements to ensure compliance and protect user data.
-
-For a comprehensive understanding, you can also read the GDPR document at [GDPR Document](/gdpr).
-
-## Key Concepts for Developers
-
-### Personal Data: Broad Scope
-
-Personal data encompasses any information that can identify an individual, directly or indirectly. This includes:
-
-- **IP addresses**: Considered personal data
-- **Device identifiers and cookies**: Classified as personal data
-- **Email addresses**: Clearly personal data
-- **User behavior data**: If it can be linked to an identifier, it's personal data
-- **Encrypted data**: Still counts if you can decrypt it
-
-### User Rights to Support
-
-Under the UK GDPR, individuals have specific rights that your applications and services must accommodate:
-
-1. **Right of Access**: Users can request access to their personal data
-   - Implement features that allow users to view and export their data in a machine-readable format
-
-2. **Right to Rectification**: Users can correct inaccurate personal data
-   - Provide functionalities for users to update their information
-
-3. **Right to Erasure**: Also known as the "right to be forgotten"
-   - Ensure mechanisms are in place to delete user data upon request, including data held in backups and third-party services
-
-4. **Right to Data Portability**: Users can obtain and reuse their personal data across different services
-   - Facilitate data export in common formats like JSON or CSV
-
-5. **Right to Object**: Users can object to the processing of their personal data
-   - Allow users to manage their consent preferences and opt out of specific data processing activities
-
-## Technical Implementation Considerations
-
-### Data Storage
-
-- **Location**: While the UK GDPR doesn't mandate data to be stored within the UK, if you transfer personal data outside the UK, ensure appropriate safeguards are in place
-- **Documentation**: Maintain records of where personal data is stored and processed
-- **Data Minimization**: Collect only the data necessary for your purposes and avoid excessive data collection
-
-### Security Measures
-
-- **Encryption**: Use encryption to protect personal data both at rest and in transit
-- **Access Controls**: Implement strict access controls to ensure only authorized personnel can access personal data
-- **Regular Audits**: Conduct regular security assessments to identify and mitigate vulnerabilities
-- **Incident Response Plan**: Develop and maintain a plan to respond to data breaches, including notifying the Information Commissioner's Office (ICO) within 72 hours if required
-
-### Cookies and Tracking Technologies
-
-- **Consent**: Obtain explicit consent before placing non-essential cookies or similar tracking technologies on users' devices
-- **Transparency**: Provide clear information about the types of cookies used and their purposes
-- **Control**: Allow users to manage their cookie preferences and withdraw consent easily
-
-## Practical Steps for Compliance
-
-1. **Data Mapping**
-   - Identify and document all personal data collected, processed, and stored by your organization
-   - Understand data flows, including transfers to third parties
-
-2. **Policy Development**
-   - Create clear privacy policies that explain your data processing activities and user rights
-   - Ensure terms of service are transparent and user-friendly
-
-3. **Training and Awareness**
-   - Educate your team about data protection principles and their responsibilities under the UK GDPR
-
-4. **Regular Reviews**
-   - Periodically assess your data protection practices and update them as necessary to maintain compliance
-
-## Key Differences Between UK GDPR and EU GDPR
-
-While the UK GDPR closely mirrors the EU GDPR, there are notable differences:
-
-- **Supervisory Authority**: In the UK, the Information Commissioner's Office (ICO) serves as the supervisory authority, replacing the European Data Protection Board's role within the EU.
-
-- **Data Transfers**: The UK is now considered a third country under the EU GDPR, necessitating additional safeguards for data transfers between the UK and EU.
-
-- **Age of Consent**: The UK GDPR sets the age of consent for data processing at 13, whereas the EU GDPR sets it at 16.
-
-- **Fines and Penalties**: While both regulations impose significant fines for non-compliance, the specific amounts and currency differ.
-
-## Resources for Developers
-
-- **Information Commissioner's Office (ICO) Guidance**: The ICO provides comprehensive resources on data protection and the UK GDPR
-  - [Guide to the UK General Data Protection Regulation (UK GDPR)](https://ico.org.uk/for-organisations/guide-to-data-protection/guide-to-the-general-data-protection-regulation-uk-gdpr/)
-
-- **National Cyber Security Centre (NCSC) Guidance**: Offers advice on implementing effective security measures
-  - [NCSC Guidance on Cloud Security](https://www.ncsc.gov.uk/collection/cloud-security)
-
-- **UK Government Digital Service (GDS) Design Principles**: While not exclusively about privacy, these principles emphasize user-centric design, which can be adapted to prioritize privacy considerations
-  - [GDS Design Principles](https://www.gov.uk/guidance/government-design-principles)
