--- apps/docs/src/content/privacy-regulations/faq.mdx
@@ -1,219 +1,0 @@
----
-title: Privacy Compliance FAQ
----
-
-## Common Technical Questions
-
-### How do we handle user data deletion requests?
-Implement a complete deletion process that:
-- Removes user data from primary databases
-- Handles cascade deletions
-- Cleans up backups
-- Documents deletion process
-- Notifies third-party services
-
-### What's required for data encryption?
-Minimum requirements include:
-- TLS for data in transit
-- AES-256 for data at rest
-- Key management system
-- Encrypted backups
-- Regular security updates
-
-### How should we handle user consent?
-Implement a consent system that:
-- Captures explicit consent
-- Records consent timestamp
-- Stores consent records
-- Allows consent withdrawal
-- Updates preferences immediately
-
-### What logging is required for compliance?
-Essential logging includes:
-- Access logs
-- Security events
-- User actions
-- System changes
-- Error conditions
-
-## Architecture Questions
-
-### How do we implement privacy by design?
-Key considerations:
-- Data minimization
-- Security by default
-- Privacy controls
-- Access management
-- Regular assessments
-
-### What's required for international data transfers?
-Requirements include:
-- Data transfer agreements
-- Appropriate safeguards
-- Documentation
-- User notification
-- Regular review
-
-### How do we handle data retention?
-Implement a retention system with:
-- Clear retention periods
-- Automated cleanup
-- Backup management
-- Documentation
-- Legal compliance
-
-## Security Questions
-
-### What are the minimum security requirements?
-Basic requirements:
-- Strong encryption
-- Access controls
-- Regular updates
-- Security monitoring
-- Incident response
-
-### How do we handle security breaches?
-Have a plan that includes:
-- Detection systems
-- Response procedures
-- Communication plans
-- Documentation
-- Recovery processes
-
-### What authentication is required?
-Implement:
-- Strong passwords
-- Multi-factor authentication
-- Session management
-- Access logging
-- Regular reviews
-
-## Compliance Questions
-
-### When do we need a DPO?
-Consider appointing a DPO when:
-- Processing sensitive data
-- Large scale processing
-- Regular monitoring
-- Public authority
-- Required by law
-
-### How often should we update privacy policies?
-Update policies when:
-- Changing data practices
-- New features launch
-- Laws change
-- New markets entered
-- Annual review
-
-### What documentation is required?
-Essential documentation:
-- Privacy policies
-- Security procedures
-- Processing records
-- Incident response
-- Training materials
-
-## Implementation Questions
-
-### How do we handle user rights requests?
-Implement processes for:
-- Request verification
-- Data retrieval
-- Response preparation
-- Documentation
-- Timeline tracking
-
-### What's required for cookie consent?
-Cookie consent needs:
-- Clear notice
-- Explicit consent
-- Preference management
-- Documentation
-- Regular updates
-
-### How do we manage third-party services?
-Required steps:
-- Vendor assessment
-- Contract review
-- Security verification
-- Regular monitoring
-- Documentation
-
-## Startup-Specific Questions
-
-### When should we start implementing privacy features?
-Start early with:
-- Basic privacy policy
-- Security foundations
-- Consent management
-- User rights handling
-- Documentation
-
-### How do we balance compliance vs. growth?
-Focus on:
-- Essential compliance
-- Scalable solutions
-- Automated processes
-- Documentation
-- Regular review
-
-### What are the minimum viable privacy requirements?
-Essential elements:
-- Privacy policy
-- Security measures
-- Consent management
-- User rights process
-- Incident response
-
-## Technical Implementation Questions
-
-### How do we handle data portability?
-Implement:
-- Data export functionality
-- Standard formats
-- Complete datasets
-- Secure transfer
-- Documentation
-
-### What's required for audit trails?
-Track and log:
-- User actions
-- System changes
-- Access attempts
-- Security events
-- Compliance activities
-
-### How do we implement data minimization?
-Strategies include:
-- Collection limits
-- Purpose specification
-- Regular review
-- Data cleanup
-- Documentation
-
-## Cost Questions
-
-### What are the essential privacy investments?
-Priority investments:
-- Security infrastructure
-- Consent management
-- Documentation systems
-- Training programs
-- Monitoring tools
-
-### How do we minimize compliance costs?
-Strategies include:
-- Open source tools
-- Automated processes
-- Clear documentation
-- Regular review
-- Team training
-
-### When should we get external help?
-Consider external help for:
-- Complex requirements
-- Legal questions
-- Security audits
-- Training programs
-- Incident response
