--- apps/docs/src/content/privacy-regulations/getting-started.mdx
@@ -1,252 +1,0 @@
----
-title: Getting Started with Privacy Compliance
----
-
-## First Steps for New Projects
-
-### 1. Initial Assessment
-- What personal data will you collect?
-- Where will users be located?
-- Which privacy laws apply?
-- What security measures are needed?
-- Which third-party services will you use?
-
-### 2. Technical Foundation
-- Secure infrastructure setup
-- Authentication system
-- Data encryption
-- Access controls
-- Logging system
-
-### 3. Documentation Basics
-- Privacy policy
-- Security procedures
-- Data inventory
-- Incident response plan
-- Training materials
-
-## Technical Implementation Guide
-
-### Data Collection
-1. **Identify Data Types**
-   - User information
-   - Usage data
-   - Technical data
-   - Location data
-   - Third-party data
-
-2. **Collection Methods**
-   - Forms
-   - APIs
-   - Cookies
-   - Analytics
-   - Third-party services
-
-3. **Storage Requirements**
-   - Encryption
-   - Access controls
-   - Backup procedures
-   - Retention periods
-   - Deletion mechanisms
-
-### User Rights Implementation
-
-1. **Access Rights**
-   - Data export functionality
-   - User authentication
-   - Data formats
-   - Response timeframes
-   - Verification process
-
-2. **Deletion Rights**
-   - Data removal process
-   - Cascade deletes
-   - Backup handling
-   - Third-party deletion
-   - Verification system
-
-3. **Other Rights**
-   - Data correction
-   - Processing restrictions
-   - Data portability
-   - Consent management
-   - Preference centers
-
-## Security Fundamentals
-
-### Basic Security Measures
-- SSL/TLS encryption
-- Secure authentication
-- Regular backups
-- Vulnerability scanning
-- Access controls
-
-### Monitoring and Logging
-- Activity logging
-- Security monitoring
-- Access tracking
-- Error logging
-- Audit trails
-
-### Incident Response
-- Detection systems
-- Response procedures
-- Communication plans
-- Documentation requirements
-- Recovery processes
-
-## Privacy by Design Implementation
-
-### Architecture Planning
-- Data minimization
-- Security by default
-- Privacy controls
-- Access management
-- Audit capabilities
-
-### Development Practices
-- Secure coding
-- Regular testing
-- Code review
-- Documentation
-- Version control
-
-### Deployment Considerations
-- Security scanning
-- Performance testing
-- Monitoring setup
-- Backup verification
-- Incident response
-
-## Compliance Automation
-
-### Key Areas to Automate
-- Consent management
-- User rights requests
-- Data deletion
-- Access controls
-- Compliance monitoring
-
-### Tools and Services
-- Privacy platforms
-- Security tools
-- Monitoring systems
-- Documentation tools
-- Training platforms
-
-## Common Implementation Challenges
-
-### Technical Issues
-1. **Data Management**
-   - Complex data structures
-   - Legacy systems
-   - Third-party integration
-   - Data synchronization
-   - Backup management
-
-2. **Security Implementation**
-   - Authentication complexity
-   - Encryption overhead
-   - Access control granularity
-   - Monitoring scale
-   - Incident detection
-
-3. **Performance Impact**
-   - Processing overhead
-   - Storage requirements
-   - Network latency
-   - Backup windows
-   - Response times
-
-### Process Issues
-1. **Documentation**
-   - Keeping records updated
-   - Version control
-   - Access management
-   - Change tracking
-   - Audit trails
-
-2. **Training**
-   - Team education
-   - Role-specific training
-   - Documentation
-   - Verification
-   - Updates
-
-## Scaling Considerations
-
-### Early Stage
-- Essential compliance
-- Basic security
-- Simple processes
-- Core documentation
-- Key monitoring
-
-### Growth Phase
-- Enhanced security
-- Automated processes
-- Regular audits
-- Comprehensive documentation
-- Team training
-
-### Enterprise Level
-- Advanced security
-- Global compliance
-- Regular assessments
-- Detailed documentation
-- Privacy team
-
-## Cost-Effective Compliance
-
-### Free and Open Source Tools
-- Security scanners
-- Monitoring tools
-- Documentation systems
-- Training resources
-- Compliance checkers
-
-### Essential Services
-- Cloud providers
-- Security services
-- Compliance platforms
-- Training providers
-- Consulting services
-
-## Next Steps
-
-### Immediate Actions
-1. Privacy policy creation
-2. Security implementation
-3. User rights process
-4. Documentation setup
-5. Team training
-
-### Short-term Goals
-1. Enhanced security
-2. Automated compliance
-3. Regular audits
-4. Comprehensive documentation
-5. Process improvements
-
-### Long-term Planning
-1. Advanced security
-2. Global compliance
-3. Regular assessments
-4. Privacy program
-5. Team expansion
-
-## Resources for Developers
-
-### Technical Resources
-- Code examples
-- Security guides
-- API documentation
-- Implementation patterns
-- Best practices
-
-### Learning Resources
-- Online courses
-- Documentation
-- Community forums
-- Webinars
-- Case studies
