--- apps/docs/src/content/privacy-regulations/meta.json
@@ -1,30 +1,0 @@
-{
-	"title": "Privacy Regulations",
-	"pages": [
-		"---Introduction---",
-		"index",
-		"privacy-compliance-best-practices",
-		"faq",
-		"---United States---",
-		"usa/ccpa",
-		"usa/cpra",
-		"usa/cpa",
-		"usa/ctdpa",
-		"usa/ucpa",
-		"usa/vcdpa",
-		"---Canada---",
-		"canada/pipeda",
-		"canada/quebec-law-25",
-		"---Europe---",
-		"europe/gdpr",
-		"europe/gdpr-uk",
-		"---Other Countries---",
-		"australia/australian-privacy-act",
-		"asia/singapore-pdpa",
-		"asia/thailand-pdpa",
-		"oceania/australia/australian-privacy-act",
-		"south-america/lgpd",
-		"middle-east/pdpl",
-		"africa/popia"
-	]
-}
