--- apps/docs/src/content/privacy-regulations/middle-east/pdpl.mdx
@@ -1,83 +1,0 @@
----
-title: Saudi Arabia's Personal Data Protection Law (PDPL)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Saudi Arabia.
----
-
-## Introduction
-
-In the evolving digital landscape, safeguarding personal information is paramount. For startups operating in or engaging with individuals in Saudi Arabia, understanding and complying with the **Personal Data Protection Law (PDPL)** is essential. Enacted to protect individual privacy rights, the PDPL outlines comprehensive requirements for handling personal data within the Kingdom.
-
-## Applicability of the PDPL
-
-The PDPL applies to any entity that processes personal data related to individuals residing in Saudi Arabia, regardless of the entity's location. This includes both public and private organizations, as well as entities outside the Kingdom that process personal data of individuals within Saudi Arabia.
-
-## Key Obligations Under the PDPL
-
-The PDPL establishes several obligations for organizations handling personal data:
-
-### 1. Lawful Basis for Processing
-
-**Requirement:** Ensure that personal data is processed based on a lawful basis, such as the individual's consent, a legal obligation, or legitimate interest.
-
-**Action Steps:**
-- **Obtain Consent:** Secure explicit consent from individuals before collecting or processing their personal data, unless another lawful basis applies.
-- **Document Legal Grounds:** Maintain records detailing the legal basis for all data processing activities.
-
-### 2. Data Subject Rights
-
-**Requirement:** Respect and facilitate individuals' rights regarding their personal data.
-
-**Action Steps:**
-- **Right to Access:** Provide individuals with access to their personal data upon request.
-- **Right to Correction:** Allow individuals to correct inaccurate or incomplete data.
-- **Right to Deletion:** Enable individuals to request the deletion of their personal data under certain conditions.
-- **Right to Object:** Honor individuals' objections to the processing of their personal data in specific scenarios.
-
-### 3. Data Security Measures
-
-**Requirement:** Implement appropriate technical and organizational measures to protect personal data from unauthorized access, disclosure, alteration, or destruction.
-
-**Action Steps:**
-- **Data Encryption:** Use encryption methods to safeguard data both in transit and at rest.
-- **Access Controls:** Restrict data access to authorized personnel only.
-- **Regular Audits:** Conduct periodic security assessments to identify and mitigate vulnerabilities.
-
-### 4. Cross-Border Data Transfers
-
-**Requirement:** Personal data transfers outside Saudi Arabia are restricted and require specific conditions to be met.
-
-**Action Steps:**
-- **Obtain Authorization:** Secure authorization from the competent authority before transferring personal data abroad.
-- **Ensure Adequate Protection:** Confirm that the recipient country provides an adequate level of data protection or implement appropriate safeguards.
-
-### 5. Data Breach Notification
-
-**Requirement:** Notify the competent authority and affected individuals of any data breaches that may cause harm.
-
-**Action Steps:**
-- **Incident Response Plan:** Develop and maintain a data breach response plan outlining notification procedures.
-- **Timely Reporting:** Report breaches promptly, adhering to the timelines specified by the PDPL.
-
-## Compliance Roadmap for Startups
-
-**Phase 1: Assessment**
-- **Data Inventory:** Compile a comprehensive inventory of personal data collected, processed, and stored by your startup.
-- **Gap Analysis:** Evaluate current data handling practices against PDPL requirements to identify compliance gaps.
-
-**Phase 2: Implementation**
-- **Policy Development:** Establish data protection policies and procedures addressing identified gaps.
-- **Employee Training:** Educate staff on PDPL obligations and best practices for data protection.
-
-**Phase 3: Monitoring and Maintenance**
-- **Regular Audits:** Conduct ongoing audits to ensure continued compliance with the PDPL.
-- **Policy Updates:** Revise policies and procedures as needed to reflect changes in the regulatory environment or business operations.
-
-## Penalties for Non-Compliance
-
-Non-compliance with the PDPL can result in significant penalties, including fines up to SAR 5 million, suspension of data processing activities, and potential reputational damage.
-
-## Conclusion
-
-Adhering to Saudi Arabia's Personal Data Protection Law is not only a legal obligation but also an opportunity to build trust with your customers by demonstrating a commitment to protecting their personal information. By proactively addressing these requirements, your startup can mitigate risks and establish a strong foundation for responsible data management.
-
-For detailed guidance and resources, consider consulting with privacy experts or utilizing compliance tools designed to assist businesses in navigating the PDPL's requirements.
