--- apps/docs/src/content/privacy-regulations/privacy-compliance-best-practices.mdx
@@ -1,52 +1,0 @@
----
-title: Privacy Compliance Best Practices
----
-
-## Data Collection Minimization
-
-### Collection Practices
-- Collect only necessary information
-- Regular data audits
-- Clear purpose documentation
-- Review collection methods regularly
-
-### Consent Management
-- Clear consent mechanisms
-- Record keeping of consent
-- Easy withdrawal options
-- Regular consent review
-
-### Documentation
-- Data inventory maintenance
-- Processing records
-- Impact assessments
-- Regular updates
-
-## Data Protection
-
-### Security Measures
-- Access controls
-- Data encryption
-- Regular security updates
-- Monitoring and logging
-
-### Staff Training
-- Regular privacy training
-- Security awareness
-- Incident response procedures
-- Role-specific training
-
-### Third-Party Management
-- Vendor assessments
-- Contractual safeguards
-- Regular compliance reviews
-- Vendor monitoring
-
-## Key Principles
-
-- Privacy by Default
-- Security by Design
-- Regular Assessment
-- Clear Documentation
-- Staff Awareness
-- Vendor Management
