--- apps/docs/src/content/privacy-regulations/south-america/lgpd.mdx
@@ -1,30 +1,0 @@
----
-title: Brazil's Lei Geral de Proteção de Dados (LGPD)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of individuals in Brazil.
----
-
-## Overview of the LGPD
-
-Brazil's Lei Geral de Proteção de Dados Pessoais (LGPD), or General Data Protection Law, is a comprehensive legal framework that regulates the processing of personal data of individuals.nacted in August 2018, the LGPD establishes guidelines for the collection, use, processing, and storage of personal data by organizations, aiming to protect individuals' fundamental rights of freedom and privacy.citeturn0search2
-## Applicability of the LGPD
-
-The LGPD applies to any individual or organization, regardless of their location, that processes personal data:
-- *In Brazil**;
-- *Of individuals located in Brazil**; or
-- *Where the processing activities are aimed at offering goods or services to individuals in Brazil**.
-his broad scope ensures that both domestic and international entities handling personal data related to individuals in Brazil are subject to the law.
-## Key Principles of Data Processing
-
-The LGPD outlines ten fundamental principles that organizations must adhere to when processing personal data:
-1. **Purpose**: Processing must be carried out for legitimate, specific, and explicit purposes.
-2. **Adequacy**: Processing must be compatible with the purposes informed to the data subject.
-3. **Necessity**: Processing should be limited to the minimum necessary data to achieve its purposes.
-4. **Free Access**: Data subjects must have free and easy access to their personal data and information about its processing.
-5. **Data Quality**: Personal data must be accurate, clear, relevant, and up-to-date.
-6. **Transparency**: Clear, accurate, and easily accessible information about data processing must be provided to data subjects.
-7. **Security**: Technical and administrative measures must be implemented to protect personal data from unauthorized access and accidental or unlawful situations.
-8. **Prevention**: Measures must be adopted to prevent the occurrence of damage due to personal data processing.
-9. **Non-Discrimination**: Processing must not be carried out for unlawful or abusive discriminatory purposes.
-10. **Accountability**: Organizations must demonstrate the adoption of effective measures to comply with data protection regulations.
-
-These principles serve as the foundation for lawful personal data processing under the LGPD.
