--- apps/docs/src/content/privacy-regulations/usa/cpa.mdx
@@ -1,49 +1,0 @@
----
-title: Colorado Privacy Act (CPA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of Colorado residents.
----
-
-## Overview of the Colorado Privacy Act
-
-he Colorado Privacy Act (CPA), enacted on July 7, 2021, and effective from July 1, 2023, establishes comprehensive data privacy rights for Colorado residents and imposes specific obligations on businesses handling personal data.
-## Applicability of the CPA
-
-he CPA applies to entities that conduct business in Colorado or produce products or services intentionally targeted to Colorado residents and that either:
-- *Control or process personal data of at least 100,000 consumers per calendar year**; or
-- *Derive revenue from the sale of personal data and control or process the personal data of at least 25,000 consumers**.
-ertain entities and data types are exempt, including state and local governments, data governed by specific federal laws, and employment records.
-## Consumer Rights Under the CPA
-The CPA grants Colorado residents the following rights regarding their personal data:
-- **Right to Access**: Consumers can confirm whether a controller is processing their personal data and access that data.
-- **Right to Correct**: Consumers can correct inaccuracies in their personal data.
-- **Right to Delete**: Consumers can request the deletion of their personal data.
-- **Right to Data Portability**: Consumers can obtain their personal data in a portable and readily usable format.
-- **Right to Opt-Out**: Consumers can opt out of the processing of their personal data for purposes of targeted advertising, the sale of personal data, or profiling in furtherance of decisions that produce legal or similarly significant effects.
-
-## Obligations for Businesses
-
-Businesses subject to the CPA must:
-- **Transparency**: Provide clear and accessible privacy notices detailing data collection and processing practices.
-- **Purpose Specification**: Collect personal data for specified, explicit, and legitimate purposes and not process it in a manner incompatible with those purposes.
-- **Data Minimization**: Limit the collection of personal data to what is adequate, relevant, and reasonably necessary for the specified purposes.
-- **Security**: Implement appropriate technical and organizational measures to secure personal data.
-- **Consent for Sensitive Data**: Obtain consumer consent before processing sensitive data, such as data revealing racial or ethnic origin, religious beliefs, health diagnoses, sexual orientation, or citizenship status.
-- **Data Protection Assessments**: Conduct assessments for processing activities that present a heightened risk of harm to consumers, including processing sensitive data and engaging in targeted advertising or profiling.
-
-## Enforcement and Penalties
-
-The Colorado Attorney General and district attorneys are responsible for enforcing the CPA. Non-compliant entities may face penalties of up to $20,000 per violation.
-
-## Preparing for CPA Compliance
-
-To comply with the CPA, businesses should:
-
-1. **Assess Applicability**: Determine if the CPA applies to your operations based on data processing activities and thresholds.
-2. **Data Mapping**: Inventory the personal data collected, processed, and shared to understand data flows and identify compliance gaps.
-3. **Update Privacy Notices**: Ensure privacy policies are transparent and include required disclosures about data practices and consumer rights.
-4. **Implement Consumer Rights Processes**: Establish procedures to respond to consumer requests to access, correct, delete, or transfer their data, and to manage opt-out preferences.
-5. **Review Data Processing Agreements**: Ensure contracts with processors include terms that require compliance with the CPA.
-6. **Conduct Data Protection Assessments**: Evaluate processing activities that may pose significant risks to consumers to identify and mitigate potential harms.
-7. **Train Employees**: Educate staff on CPA requirements and internal policies to ensure consistent compliance across the organization.
-
-By proactively addressing these areas, businesses can align with the CPA's requirements and uphold consumer privacy rights in Colorado.
