--- apps/docs/src/content/privacy-regulations/usa/cpra.mdx
@@ -1,142 +1,0 @@
----
-title: California Privacy Rights Act (CPRA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of California residents.
----
-
-## Why CPRA Matters for Your Startup
-
-If your startup collects data from California residents—especially if you handle user consent—you must comply with the **California Privacy Rights Act (CPRA)**. CPRA expands on the **California Consumer Privacy Act (CCPA)** and introduces stricter requirements for data privacy, including new rules for sensitive personal information and automated decision-making.
-
-Even if your startup doesn't meet CPRA's compliance thresholds today, implementing best practices early will help you scale efficiently and avoid legal risks as you grow.
-
-## Does CPRA Apply to Your Startup?
-
-Your startup must comply if **any** of the following apply:
-
-- **Annual revenue** exceeds $25 million.
-- **Data Transactions:** You buy, sell, or share the personal data of **100,000+ consumers/households** per year (up from CCPA’s 50,000 threshold).
-- **Revenue from Data:** At least **50% of annual revenue** comes from selling or sharing personal information.
-
-If you handle user consent and process personal data, compliance is essential—especially with stricter **opt-in and opt-out** requirements.
-
-## What’s New in CPRA? Key Changes for Startups
-
-### 1. **Sensitive Personal Information (SPI) Protections**
-CPRA introduces a new category: **Sensitive Personal Information (SPI)**. If you collect or process SPI, you must provide users with additional rights and controls.
-
-**Examples of SPI:**
-- Social Security & ID numbers
-- Financial account information
-- Precise geolocation
-- Race, ethnicity, & religion
-- Private communications (emails, messages)
-- Genetic & biometric data
-- Health & sexual orientation data
-
-**Action Items for Your Startup:**
-- Allow users to **limit** how their SPI is used.
-- Update **privacy policies** to disclose SPI collection and usage.
-- Implement **data minimization**—collect only what's necessary.
-
-### 2. **Stronger User Rights & Consent Management**
-
-#### **Right to Correction**
-- Users can request **corrections** to their data.
-- Startups must verify and **document** corrections.
-- Notify **service providers** (third parties) of corrections.
-
-#### **Right to Limit Use of SPI**
-- Users can **restrict processing** of SPI.
-- Implement an easy-to-use **consent management** tool.
-- Honor **global privacy signals** indicating user preferences.
-
-#### **Right to Opt-Out of Automated Decision Making**
-- If your startup uses algorithms for personalization, profiling, or AI-based decision-making, **users can opt out**.
-- Offer a way for users to **request human review** of decisions.
-- Provide **transparent disclosures** about automated processes.
-
-#### **Enhanced Right to Delete**
-- Startups must **delete personal data** upon request.
-- Notify **third parties** to delete shared data.
-- Maintain a **deletion log** to track compliance.
-
-#### **Right to Non-Discrimination**
-- Users cannot be penalized for exercising privacy rights.
-- If you offer **incentives for data collection**, they must be **transparent and fair**.
-
-### 3. **Stronger Security & Data Management Requirements**
-
-#### **Data Minimization & Retention Limits**
-- Collect only the **minimum data** necessary for business purposes.
-- Define and document **data retention policies**.
-- Implement **automatic deletion** for expired data.
-
-#### **Stronger Security Measures**
-- Conduct **risk assessments** to evaluate privacy risks.
-- Implement **encryption, access controls, and monitoring**.
-- Prepare **incident response plans** for data breaches.
-
-#### **Cross-Border Data Transfers**
-- If you transfer data internationally, ensure **adequate protections**.
-- Document **legal transfer mechanisms** (e.g., Standard Contractual Clauses).
-
-## How to Implement CPRA Compliance for Your Startup
-
-### 1. **Upgrade Your Consent Management System**
-- Support **granular consent preferences** (e.g., SPI vs. general data).
-- Implement **automated compliance tracking** for consent changes.
-- Ensure compliance with **Do Not Sell/Share** requests.
-
-### 2. **Enhance Your Privacy Policy & Notices**
-- Clearly disclose **what data you collect and why**.
-- Update policies to reflect **SPI protections** and **user rights**.
-- Include a **“Limit My SPI Use”** option in your interface.
-
-### 3. **Develop Efficient User Request Handling**
-- Automate **data access, correction, deletion, and opt-out** processes.
-- Verify and fulfill requests within **45 days**.
-- Notify **third-party partners** of any data deletions or corrections.
-
-### 4. **Implement Stronger Security & Compliance Monitoring**
-- Regularly conduct **security audits** and **compliance reviews**.
-- Encrypt **SPI and other sensitive data**.
-- Maintain **detailed compliance records**.
-
-### 5. **Educate Your Team**
-- Train developers on **privacy-by-design principles**.
-- Ensure all employees understand **CPRA obligations**.
-- Assign a **privacy lead** to oversee compliance efforts.
-
-## Key Technical Challenges & Solutions
-
-### 1. **Managing Sensitive Data Compliance**
-- Use **automated classification tools** to flag SPI.
-- Implement **role-based access controls** for SPI.
-- Set up **automated retention and deletion policies**.
-
-### 2. **Handling Automated Decision-Making Requests**
-- Maintain **clear documentation** of algorithm logic.
-- Provide **human review options** where required.
-- Track **decision-making impact assessments**.
-
-### 3. **Verifying Consumer Requests Securely**
-- Implement **multi-factor authentication (MFA)** for identity verification.
-- Log and **track requests** to ensure compliance.
-- Maintain **audit records** of user interactions.
-
-## Compliance Timeline & Next Steps
-
-- **January 1, 2023** – CPRA takes effect
-- **July 1, 2023** – Enforcement begins
-- **January 1, 2024** – Updated contracts required for third parties
-- **Ongoing** – Regular compliance reviews & updates
-
-## Final Thoughts: How to Stay Ahead
-
-### **Best Practices for CPRA Compliance in Startups**
-- Implement **privacy-by-design** in your product development.
-- Automate **consent and user rights management**.
-- Monitor **regulatory updates** to stay compliant.
-- Prioritize **user transparency and trust** in data handling.
-
-By embedding **CPRA compliance into your startup's consent management strategy**, you’ll not only meet legal requirements but also build a foundation of trust with your users.
