--- apps/docs/src/content/privacy-regulations/usa/ctdpa.mdx
@@ -1,79 +1,0 @@
----
-title: Connecticut Data Privacy Act (CTDPA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of Connecticut residents.
----
-
-## Why You Need to Care About CTDPA
-
-The Connecticut Data Privacy Act (CTDPA), effective from July 1, 2023, establishes comprehensive data privacy rights for Connecticut residents and imposes specific obligations on businesses handling personal data. Understanding and complying with the CTDPA is crucial for startups operating in or targeting consumers in Connecticut.
-
-## When CTDPA Applies to Your Startup
-
-The CTDPA applies to entities conducting business in Connecticut or producing products or services targeted to Connecticut residents, and that meet one of the following criteria:
-
-- **Control or process the personal data of at least 100,000 consumers annually**, excluding data processed solely for payment transactions.
-- **Derive over 25% of gross revenue from the sale of personal data** and process the personal data of at least 25,000 consumers.
-
-Notably, the CTDPA does not include a revenue threshold, focusing instead on the volume of data processed and the nature of the business.
-
-## What Counts as Personal Data
-
-Under the CTDPA, "personal data" refers to any information linked or reasonably linkable to an identified or identifiable individual. This excludes de-identified data or publicly available information.
-
-The act also designates certain information as "sensitive data," including:
-- Data revealing racial or ethnic origin
-- Religious beliefs
-- Mental or physical health conditions
-- Sexual orientation
-- Citizenship or immigration status
-- Genetic or biometric data
-- Children's data
-- Precise geolocation data
-
-## Key Rights You Need to Support
-
-The CTDPA grants consumers several rights concerning their personal data:
-
-### Right to Access
-
-Consumers can:
-- Confirm whether a controller is processing their personal data
-- Access their personal data
-- Receive a response within 45 days (with possible 45-day extension)
-
-### Right to Correct
-
-Consumers can request corrections to inaccuracies in their personal data. Your systems should:
-- Provide a clear correction request process
-- Verify requester identity
-- Update information across all platforms and backups
-- Notify service providers to update the information
-
-### Right to Delete
-
-Consumers can request deletion of their personal data. Your systems must:
-- Implement a straightforward deletion request process
-- Verify requester identity
-- Ensure complete deletion across all systems
-- Inform service providers to delete the data
-
-### Right to Data Portability
-
-Consumers can:
-- Obtain their personal data in a portable format
-- Receive data in a readily usable format when technically feasible
-- Transfer their data to other services
-
-### Right to Opt-Out
-
-Consumers can opt out of:
-- Processing for targeted advertising
-- Sale of personal data
-- Profiling for decisions that produce legal or similarly significant effects
-
-Your business must:
-- Provide clear opt-out mechanisms
-- Process opt-out requests promptly
-- Honor user preferences consistently
-- Inform third parties of opt-out status
-- Avoid discrimination against users who exercise these rights
