--- apps/docs/src/content/privacy-regulations/usa/ucpa.mdx
@@ -1,139 +1,0 @@
----
-title: Utah Consumer Privacy Act (UCPA)
-description: A comprehensive data protection law that regulates how businesses collect, process, and protect personal data of Utah residents.
----
-## Why UCPA Matters for Your Startup
-
-If your startup collects data from Utah residents, understanding the Utah Consumer Privacy Act (UCPA) is crucial. While UCPA is less restrictive than CCPA and CPRA, it still introduces key privacy obligations that developers and businesses must follow.
-
-## When UCPA Applies to Your Business
-
-UCPA applies if your startup:
-- **Has annual revenue of $25 million or more** AND **either**:
-  - Controls or processes personal data of **100,000+ Utah consumers annually**
-  - Derives **50%+ revenue from selling personal data** while processing **25,000+ consumers' data**
-
-Even if your startup doesn't meet these thresholds, integrating privacy protections early will help future-proof your compliance strategy.
-
-## What Personal Data Is Covered?
-
-UCPA covers a broad range of **personal data**, including:
-- **Traditional Identifiers**: Name, email, phone number
-- **Technical Data**: IP addresses, device IDs, cookies
-- **Behavioral Data**: Browsing history, purchase history
-- **Geolocation Data**: Precise or approximate location
-- **Employment & Education Data**: Job titles, academic records
-- **Inferences**: User profiles derived from other data points
-
-**Exclusions:** Unlike stricter laws like CPRA, UCPA **excludes** publicly available data, anonymized data, and certain employment-related information.
-
-## Key Consumer Rights & Compliance Requirements
-
-### Right to Know
-Consumers can request:
-- The **categories of personal data** collected
-- The **purposes** for data processing
-- The **third parties** data is shared with
-- **Retention periods** for personal data
-
-### Right to Delete
-Consumers can request deletion of their personal data **if provided directly to your business**. Your startup must:
-- Offer a **clear deletion mechanism**
-- **Verify** identity before processing deletion requests
-- Ensure data is removed from **all systems and backups**
-- Notify third-party service providers to delete the data
-- Maintain **audit logs** of deletion requests
-
-### Right to Data Portability
-Consumers can request a **readily usable format** of their data. Your startup should:
-- Provide **structured, machine-readable** exports
-- Include **all user-submitted personal data**
-- Ensure **secure** transfer methods
-
-### Right to Opt-Out of Data Sales & Targeted Advertising
-Consumers have the right to opt out of:
-- **Sale of personal data**
-- **Targeted advertising based on personal data**
-- **Profiling in automated decision-making**
-
-Your business must:
-- Provide a **Do Not Sell My Personal Data** link (if applicable)
-- Support **opt-out preference signals**
-- Honor requests **without requiring user accounts**
-- Process opt-outs **within a reasonable timeframe**
-
-### No Private Right of Action
-Unlike California's CCPA, UCPA **does not allow individuals to sue businesses** for violations. Only the Utah Attorney General can enforce compliance.
-
-## Technical Implementation Guide
-
-### 1. Data Inventory & Mapping
-- **Track** all personal data collected
-- **Document** sources, processing purposes, and sharing practices
-- **Map data flows** between internal systems and third parties
-- **Set clear retention policies** for different data categories
-
-### 2. Identity Verification
-- Implement **secure identity verification** for data requests
-- Use **multi-factor authentication** where possible
-- Support **requests from authorized agents**
-- Maintain **detailed verification logs**
-
-### 3. Consumer Request Handling
-- Process **requests within 45 days** (one 45-day extension allowed)
-- Maintain **clear documentation** of all requests
-- Use **secure communication methods** for data transfers
-- Standardize responses to ensure compliance
-
-### 4. Opt-Out & Consent Management
-- Support **global opt-out signals**
-- Provide **simple UI options** for opt-out requests
-- Regularly update **preference management tools**
-
-### 5. Security & Data Protection
-- Encrypt **sensitive personal data** at rest and in transit
-- Implement **access controls** for internal data handling
-- Monitor for **unauthorized access** or data breaches
-- Have a **breach response plan** in place
-
-## Common Startup Challenges & Solutions
-
-### **Challenge: Identifying All Data Sources**
-✅ **Solution:** Conduct **regular data audits** to track new data sources.
-
-### **Challenge: Handling Third-Party Data**
-✅ **Solution:** **Review contracts** with vendors and ensure they comply with UCPA.
-
-### **Challenge: Implementing a Scalable Compliance Process**
-✅ **Solution:** Use **privacy automation tools** to streamline compliance tasks.
-
-## UCPA vs. Other Privacy Laws
-| Feature | UCPA | CCPA/CPRA | GDPR |
-|---------|------|----------|------|
-| Revenue Threshold | $25M+ | $25M+ | Any |
-| Consumer Rights | Limited | Broad | Extensive |
-| Private Right of Action | ❌ No | ✅ Yes | ✅ Yes |
-| Sensitive Data Category | ❌ No | ✅ Yes | ✅ Yes |
-| Global Opt-Out Support | ✅ Yes | ✅ Yes | ✅ Yes |
-| Enforcement | Attorney General | Privacy Agency + AG | Data Protection Authorities |
-
-## Compliance Roadmap for Startups
-
-**Phase 1: Initial Setup**
-- Identify **personal data collection points**
-- Update **privacy policies** and disclosures
-- Establish a **data subject request system**
-
-**Phase 2: Process Implementation**
-- Develop **deletion and portability workflows**
-- Train **staff on privacy requirements**
-- Review **vendor contracts** for compliance
-
-**Phase 3: Ongoing Compliance**
-- Conduct **annual audits** of data practices
-- Monitor **privacy trends** and legal updates
-- Automate **compliance reporting** where possible
-
-## Final Thoughts
-UCPA compliance is **simpler than CCPA/GDPR**, but startups should still prioritize **data transparency, user rights, and security**. By embedding privacy into your processes early, you can avoid costly changes down the line and build trust with your users.
-
