--- apps/docs/src/content/release-notes/1.0.0-rc.1.mdx
@@ -1,9 +1,0 @@
----
-title: 1.0.0-rc.1
-description: we are releasing the first beta version of the c15t.com framework
----
-
-## 1.0.0-rc.1
-
-- Initial release
-
