--- apps/docs/src/content/release-notes/index.mdx
@@ -1,9 +1,0 @@
----
-title: Release Notes
-description: c15t.com transforms privacy consent from a compliance checkbox into a fully observable system. Built for modern development teams, it unifies analytics, consent tracking, and privacy controls into a single performant solution - no more slow cookie banners or blind spots in user privacy choices.
----
-
-## 0.0.1
-
-- Initial release
-
