--- apps/docs/src/content/release-notes/meta.json
@@ -1,4 +1,0 @@
-{
-	"title": "Release Notes",
-	"pages": ["release-notes", "index", "---Release Notes---", "1.0.0-rc.1"]
-}
